Show promote dialog only on valid move
With a game with black king on B8 and white pawn on B7. When selecting
white pawn to go over the king I could have the promote dialog, after
selecting queen for example the game go back to the original state.

This patch check that the move is valid first before showing the promote
dialog.